update 'put variation' verbiage to plural in 'test_projection_protein_edges'
