+++ tests/integration/restapi/variations_router/test_variation.py
@@ -207,6 +207,16 @@ def test_post_variation_invalid_request(
     assert resp.json() == {"detail": err_msg}
 
 
+def test_get_allele(restapi_client: TestClient, preloaded_alleles: dict[str, str]):
+    for allele_id, allele_fixture in preloaded_alleles.items():
+        resp = restapi_client.get(f"/variations/{allele_id}")
+        assert resp.status_code == HTTPStatus.OK
+        assert resp.json()["data"] == allele_fixture["variation"]
+
+    bad_resp = restapi_client.get("/object/ga4gh:VA.invalid7DSM9KE3Z0LntAukLqm0K2ENn")
+    assert bad_resp.status_code == HTTPStatus.NOT_FOUND
+
+
 def test_simple_delete_object(restapi_client: TestClient, alleles: dict):
     for allele_fixture in alleles.values():
         restapi_client.put("/vrs_variation", json=allele_fixture["variation"])
